+++ src/app/member/member.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import {Observable} from 'rxjs';
 import {default as _members} from '../../assets/members.json';
+import {HttpClient} from '@angular/common/http';
 
 @Component({
   selector: 'app-member',
@@ -37,7 +38,7 @@ export class MemberComponent implements OnInit {
       abbr: 'SD'
     }];
 
-  constructor(private route: ActivatedRoute) {}
+  constructor(private route: ActivatedRoute, private http: HttpClient) {}
 
   ngOnInit() {
     this.id = this.route.snapshot.params.id;
@@ -48,10 +49,24 @@ export class MemberComponent implements OnInit {
       const votes = this.member.aggregate_votes[voteType];
       this.member['perc' + voteType] = (votes / totalVotes * 100).toFixed(1);
     });
+    this.getMemberPhoto(this.id).then(url => {
+      this.member.imageURL = url;
+    });
   }
 
   getParty(abbr) {
     return this.parties.filter(pair => pair.abbr === abbr)[0].name;
   }
 
+  async getMemberPhoto(id) {
+    console.log(id);
+    return new Promise((resolve, reject) => {
+      this.http.get(`http://data.riksdagen.se/personlista/?iid=${id}`, {responseType: 'text'}).subscribe((data) => {
+        const re = /http:\/\/data.riksdagen.se\/filarkiv\/bilder\/ledamot\/.*_max.jpg/;
+        const url = data.match(re)[0];
+        console.log(url);
+        resolve(url);
+      });
+    });
+  }
 }
