+++ src/app/memberList/memberList.component.ts
@@ -29,23 +29,12 @@ export class MemberListComponent implements OnInit {
 
   @ViewChild(MatSort) sort: MatSort;
 
-  constructor(private http: HttpClient) {}
+  constructor() {}
 
   ngOnInit() {
     console.log(this.members);
     this.dataSource.sort = this.sort;
-    this.getMemberPhoto(this.members[0].id);
   }
 
-  async getMemberPhoto(id) {
-    console.log(id);
-    return new Promise((resolve, reject) => {
-      this.http.get(`http://data.riksdagen.se/personlista/?iid=${id}`, {responseType: 'text'}).subscribe((data) => {
-        const re = /http:\/\/data.riksdagen.se\/filarkiv\/bilder\/ledamot\/.*_192.jpg/;
-        const url = data.match(re)[0];
-        console.log(url);
-        resolve(url);
-      });
-    });
-  }
+
 }
